Fixed a crash bug where we search a user, tap one of them, get their followers and tap any user.
Reproduce steps and explanation:
1- search for a name
2- use the search feature to eliminate a couple of people
3- tap one of them
4- get followers
5- expected, search should be applied to the new followers or search cleaned
6- tap one of the people and the app crashes

The old search and cells remain but followers and filteredFollowers are empty. isSearching is true when get followers clicked and we dismiss the screen, so when we click we get the filteredFollowers(which is empty), and with the tapping, it's gonna crash.

To fix that, I basically move the searchController to the class level and removed the text when the user taps the get followers.